// state/Cart/cart.action-types.ts
export enum ActionTypes {
  ADD_CART_ITEM_START = 'ADD_CART_ITEM_START',
  ADD_CART_ITEM_SUCCESS = 'ADD_CART_ITEM_SUCCESS',
  ADD_CART_ITEM_ERROR = 'ADD_CART_ITEM_ERROR',

  GET_CART_START = 'GET_CART_START',
  GET_CART_SUCCESS = 'GET_CART_SUCCESS',
  GET_CART_ERROR = 'GET_CART_ERROR',

  CALCULATE_PRICES = 'CALCULATE_PRICES',
  REMOVE_CART_ITEM = 'REMOVE_CART_ITEM',
  SAVE_CART_SHIPPING_ADDRESS = 'SAVE_CART_SHIPPING_ADDRESS',
  SAVE_CART_PAYMENT_METHOD = 'SAVE_CART_PAYMENT_METHOD',

  CLEAR_CART = 'CLEAR_CART',

}